Zaktrak-backend/src/main/java/com/ZakTrak/service/UserService.java
package com.ZakTrak.service;
import com.ZakTrak.model.User;
import com.ZakTrak.repository.UserRepository;
import org.springframework.security.crypto.password.PasswordEncoder;
import org.springframework.stereotype.Service;
import com.ZakTrak.config.SecurityConfig;

@Service
public class UserService {

    private final UserRepository userRepository;
    private final PasswordEncoder passwordEncoder;

    public UserService(UserRepository userRepository, PasswordEncoder passwordEncoder) {
        this.userRepository = userRepository;
        this.passwordEncoder = passwordEncoder;
    }

    public User createUser(String email, String password) {
        if (userRepository.findByEmail(email) != null) {
            throw new IllegalArgumentException("Email already exists");
        }

        String encodedPassword = passwordEncoder.encode(password);
        System.out.println(encodedPassword);
        User user = new User(email, encodedPassword);
        return userRepository.save(user);
    }

}
